Hide the ugly stuff and cleanup unused parts
Refactor of Flask stuff to its own file.  The more complex Reddit api
calls are now "hidden" in the backend program.  This will help, as it
will really get complex from here.

The save_created_state was an example from the reddit api, not necessary
at this time.  Obviously wasn't doing much.
